+++ src/settings.ts
@@ -1,37 +1,14 @@
 import {App, PluginSettingTab, Setting} from "obsidian";
 import PlutoHubPlugin from "./main";
 import { t } from './utils/translation';
+import { PlutoSettings } from "types/pluto";
 
-// 定义模块的元数据接口
-export interface MiniModule {
-    id: string;
-    name: string;
-    enabled: boolean;
-    bgColor?: string; // 用于存储随机渐变色
-    bgUrl?: string; // 用于存储图片 URL
-    files: ModFile[]; // 直接在MiniModule中包含文件数组
-}
-export interface ModFile {
-    name: string;
-    type: string;
-    content: string;
-}
-
-// 定义插件设置接口
-export interface PlutoSettings {
-    moduleStoragePath: string;
-	backupFolderName: string
-    usePako: boolean;
-    columns: number;
-    modules: string[];
-}
 // 默认设置：动态获取 configDir
 export const DEFAULT_SETTINGS: PlutoSettings  = {
     moduleStoragePath: `.obsidian/cache/modules`,
 	backupFolderName: 'backups',
     usePako: true,
-    columns: 3,
-    modules: []
+    columns: 3
 };
 
 export class PlutoSettingTab extends PluginSettingTab {